Adds ICloneableConverter interface to allow for proper transfer of converters, refactor location of adapters
